+++ src/components/mod.rs
@@ -1,4 +1,5 @@
 mod asteroid;
+mod constant_orbit;
 mod engine;
 mod gate;
 mod gate_connection;
@@ -13,6 +14,7 @@ mod station;
 mod trade;
 
 pub use {
-    asteroid::*, engine::Engine, gate::*, gate_connection::*, in_sector::*, inventory::Inventory,
-    planet::*, sector::*, selectable_entity::*, ship::*, star::*, station::*, trade::*,
+    asteroid::*, constant_orbit::*, engine::Engine, gate::*, gate_connection::*, in_sector::*,
+    inventory::Inventory, planet::*, sector::*, selectable_entity::*, ship::*, star::*, station::*,
+    trade::*,
 };
