+++ src/simulation/physics/constant_velocity.rs
@@ -3,15 +3,26 @@ use bevy::prelude::{Component, Vec2};
 /// Guaranteed to never change.
 #[derive(Component, Default)]
 pub struct ConstantVelocity {
-    pub velocity: Vec2,
-    pub sprite_rotation: f32,
+    velocity: Vec2,
+    sprite_rotation: f32,
 }
 
 impl ConstantVelocity {
+    #[inline]
     pub fn new(velocity: Vec2, sprite_rotation: f32) -> ConstantVelocity {
         ConstantVelocity {
             velocity,
             sprite_rotation,
         }
+    }
+
+    #[inline]
+    pub fn velocity(&self) -> Vec2 {
+        self.velocity
+    }
+
+    #[inline]
+    pub fn sprite_rotation(&self) -> f32 {
+        self.sprite_rotation
     }
 }
