+++ src/simulation/physics/plugin.rs
@@ -39,7 +39,7 @@ fn move_constant_stuff(
     mut items: Query<(&mut SimulationTransform, &ConstantVelocity)>,
 ) {
     items.par_iter_mut().for_each(|(mut transform, velocity)| {
-        transform.rotate(velocity.sprite_rotation * time.delta_seconds());
-        transform.translation += velocity.velocity * time.delta_seconds();
+        transform.rotate(velocity.sprite_rotation() * time.delta_seconds());
+        transform.translation += velocity.velocity() * time.delta_seconds();
     });
 }
